+++ CHANGELOG.md
@@ -544,3 +544,4 @@
   - so its Item list ( worry list ) is stale
 - `doOneRound` fixed! iterates over keys, not values (Monkeys)
 - `doOneOp` changed to work with keys, not values (Monkeys)
+- `PartA` success!